add tests for post comment api endpoints, restructs tests folders for recipe and forum
